feat(PurchaseLottoService): 구매 금액 입력시 구매금액만큼 해당 로또를 생성하는지 테스트 및 기능 구현
