+++ docs/REQUIREMENTS.md
@@ -28,8 +28,8 @@
 
 - [x] 구입금액을 인자로 받는다.
 - [x] 로또 1장의 가격은 1,000원 이다.
-- [ ] 중복없이 1~45 중 6개의 숫자를 뽑아 랜덤숫자 배열을 생성한다.
-- [ ] 로또 구입 금액을 입력하면 구입 금액에 해당하는 만큼 로또를 발행해야 한다.
+- [x] 중복없이 1~45 중 6개의 숫자를 뽑아 랜덤숫자 배열을 생성한다.
+- [x] 로또 구입 금액을 입력하면 구입 금액에 해당하는 만큼 로또를 발행해야 한다.
 - [ ] 구매한 로또를 출력한다.
   - [ ] 로또 번호는 오름차순으로 정렬하여 보여준다.
